requirements.txt
pymysql==0.9.3
django==3.0.3
asgiref==3.2.3
chardet==3.0.4
defusedxml==0.6.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0 
django-formtools==2.2
django-import-export==2.0.2
django-jsonfield==1.4.0
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.17.0
idna==2.9
Pillow==7.0.0
pip==19.0.3
pytz ==2019.3
PyYAML==5.3
requests==2.23.0
six ==1.14.0
sqlparse==0.3.0
wheel==0.34.2
django-pure-pagination==0.3.0
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0

